+++ UIAdapter/capnzero/MidiEmLoader/LoaderRpc.h
@@ -20,7 +20,7 @@ public:
    void loadMusicDeviceToChain(const ::capnzero::TextView& chainRoot,
                                const ::capnzero::TextView& manufacturer,
                                const ::capnzero::TextView& device,
-                               ::capnzero::UInt8 midiVoiceOffset) override;
+                               ::capnzero::Int16 midiVoiceOffset) override;
    void removeLastMusicDeviceFromChain(
        const ::capnzero::TextView& chainRoot) override;
