+++ UIAdapter/capnzero/MidiEmRt/InstrumentsRpc.h
@@ -12,10 +12,10 @@ public:
 	InstrumentsRpc(base::instruments::Instruments& rInstruments);
 	void createKitInstrument(const ::capnzero::TextView& name) override;
 	void removeKitInstrument(const ::capnzero::SpanCL<16>& instrumentUUID) override;
-	void melodicNoteOn(const ::capnzero::SpanCL<16>& uuid, ::capnzero::UInt8 note, ::capnzero::Float32 velocity) override;
-	void melodicNoteOff(const ::capnzero::SpanCL<16>& uuid, ::capnzero::UInt8 note, ::capnzero::Float32 velocity) override;
-	void kitNoteOn(const ::capnzero::SpanCL<16>& uuid, ::capnzero::UInt8 soundIndex, ::capnzero::UInt8 note, ::capnzero::Float32 velocity) override;
-	void kitNoteOff(const ::capnzero::SpanCL<16>& uuid, ::capnzero::UInt8 soundIndex, ::capnzero::UInt8 note, ::capnzero::Float32 velocity) override;
+	void melodicNoteOn(const ::capnzero::SpanCL<16>& uuid, ::capnzero::Int16 note, ::capnzero::Float32 velocity) override;
+	void melodicNoteOff(const ::capnzero::SpanCL<16>& uuid, ::capnzero::Int16 note, ::capnzero::Float32 velocity) override;
+	void kitNoteOn(const ::capnzero::SpanCL<16>& uuid, ::capnzero::Int16 soundIndex, ::capnzero::Int16 note, ::capnzero::Float32 velocity) override;
+	void kitNoteOff(const ::capnzero::SpanCL<16>& uuid, ::capnzero::Int16 soundIndex, ::capnzero::Int16 note, ::capnzero::Float32 velocity) override;
 
 private:
 	base::instruments::Instruments& m_rInstruments;
